+++ GenericsHomework/Node.cs
@@ -1,34 +1,34 @@
 ﻿
 namespace GenericsHomework
 {
-    public class Node<T> 
+    public class Node<T>
     {
         public Node(T TValue)
         {
             this.Value = TValue;
             this.Next = this;
-            
+
         }
 
-        public T Value { get; set;}
+        public T Value { get; set; }
         public Node<T> Next { get; private set; }
 
         public override string? ToString()
         {
-            string? value = Value?.ToString(); 
+            string? value = Value?.ToString();
             return value;
         }
 
         public void Append(T TValue)
         {
-          
+
             if (this.Exists(TValue))
             {
                 throw new ArgumentException($"Duplicate Value not permitted:{nameof(TValue)}");
             }
 
-            Node<T> node = new Node<T> (TValue);
-            this.Next = node; 
+            Node<T> node = new(TValue);
+            this.Next = node;
 
 
         }
@@ -36,22 +36,22 @@ namespace GenericsHomework
         public void Clear()
         {
             var nextNode = this.Next;
-            var currentNode = this;
+            _ = this;
 
             while (nextNode != this)
             {
-               currentNode = nextNode;
-               nextNode = currentNode.Next;
-               currentNode.Next = currentNode;
+                Node<T>? currentNode = nextNode;
+                nextNode = currentNode.Next;
+                currentNode.Next = currentNode;
             }
 
-            this.Next = this; 
+            this.Next = this;
             // Given a circular list we don't need to worry about garbage collection, as we've made all Nodes past the current Node unreachable by having it refer back to itself. No doubley circular linked lists please :(
         }
 
-        public bool Exists( T TValue)
+        public bool Exists(T TValue)
         {
-            Node<T> nextNode = this.Next;
+            _ = this.Next;
             Node<T> currentNode = this;
             bool endOfList = false;
             bool exists = false;
@@ -62,11 +62,12 @@ namespace GenericsHomework
                 {
                     exists = true;
                 }
-                if (currentNode.Next == this)
+                if (currentNode.Next == currentNode)
                 {
                     endOfList = true;
                 }
-                nextNode = currentNode.Next;
+
+                currentNode = currentNode.Next;
                 // not at end of list
 
             }
